don't build rpm with GSS. do not depend on mblist-window.c Clear up the
* balsa.spec.in: don't build rpm with GSS.
* src/mailbox-conf.c: do not depend on mblist-window.c
* src/main-window.[hc]: Clear up the mailbox/message referencing code.
* src/print.c: (int) conversion in isspace()
* src/sendmsg-window.c,src/balsa-message.c:: Russian KOI and korean support
* INSTALL, README: warn about the libgnome bug and RH 7.0 gcc issues.

+++ src/print.c
@@ -314,7 +314,7 @@ print_wrap_string(gchar * str, GnomeFont * font, gint width)
 	line_width = 0;
 	last_space = NULL;
 	while (*ptr && (line_width <= width || !last_space)) {
-	    if (isspace(*ptr)) {
+	    if (isspace((int)*ptr)) {
 		*ptr = ' ';
 		last_space = ptr;
 	    }
